+++ RELEASE_NOTES.md
@@ -1,6 +1,12 @@
 ### OVERVIEW
 This module wraps the RAST annotation pipeline for KBase.
 
+### Version 0.1.6
+__Changes__
+annotate_genome(s) now takes a taxon ID and timestamp to allow exact specification of the taxon ID
+rather than submitting a scientific name. The scientific name is retrieved from the Relation
+Engine using the timestamp.
+
 ### Version 0.1.5
 __Changes__
 Prepended the scientific name display in the UI with NCBI.